third_party/WebKit/LayoutTests/http/tests/inspector/security/active-subresource-with-cert-errors.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../security-test.js"></script>
<script>
function test()
{
    /** @type {!Protocol.Security.InsecureContentStatus} */
    var insecureContentStatus = { ranMixedContent: false, displayedMixedContent: false, ranContentWithCertErrors: true, displayedContentWithCertErrors: false, ranInsecureContentStyle: Protocol.Security.SecurityState.Insecure, displayedInsecureContentStyle: Protocol.Security.SecurityState.Neutral };

    InspectorTest.mainTarget.model(Security.SecurityModel).dispatchEventToListeners(Security.SecurityModel.Events.SecurityStateChanged, new Security.PageSecurityState(Protocol.Security.SecurityState.Insecure, true, [], insecureContentStatus, null));

    var request = new SDK.NetworkRequest(0, "http://foo.test", "https://foo.test", 0, 0, null);
    InspectorTest.dispatchRequestFinished(request);

    var explanations = Security.SecurityPanel._instance()._mainView.contentElement.getElementsByClassName("security-explanation");
    for (var i = 0; i < explanations.length; i++)
        InspectorTest.dumpDeepInnerHTML(explanations[i]);
    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests addition of explanation when an active subresource was loaded with certificate errors.</p>
</body>
</html>
