third_party/WebKit/LayoutTests/http/tests/inspector/security/active-subresource-with-cert-errors-expected.txt

Tests addition of explanation when an active subresource was loaded with certificate errors.

<DIV class=security-explanation security-explanation-insecure >
    <DIV class=security-property security-property-insecure >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Active content with certificate errors
        </DIV>
        <DIV >
You have recently allowed content loaded with certificate errors (such as scripts or iframes) to run on this site.
        </DIV>
    </DIV>
</DIV>
